requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
python-multipart